rename and adding to it's own module ProgressBar that replaces StatInfos
